Reflexion-Agent/pyproject.toml
[tool.poetry]
name = "reflexion-agent"
version = "0.1.0"
description = ""
authors = ["Navin3d <smnavin65@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = {extras = ["openai"], version = "^1.0.0"}
langgraph = "^1.0.0"
langchain-ollama = "^1.0.0"
python-dotenv = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
